+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Bugs:
 
 - `[ext_jinja2]` Refactor hard-coded reference to `jinja2` template handler.
   - [Issue #749](https://github.com/datafolklabs/cement/issues/749)
+- `[ext_smtp]` Misc fixes and updates to better support content types.
+  - [PR #742](https://github.com/datafolklabs/cement/pull/742)
 
 Features: